license and move utilities
