+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/RateLimiter.java
@@ -1,7 +1,7 @@
 package io.confluent.parallelconsumer.internal;
 
 /*-
- * Copyright (C) 2020-2021 Confluent, Inc.
+ * Copyright (C) 2020-2022 Confluent, Inc.
  */
 
 import lombok.Getter;
