+++ parallel-consumer-core/src/test/java/io/confluent/csid/utils/ProgressBarUtils.java
@@ -1,7 +1,7 @@
 package io.confluent.csid.utils;
 
 /*-
- * Copyright (C) 2020-2021 Confluent, Inc.
+ * Copyright (C) 2020-2022 Confluent, Inc.
  */
 
 import lombok.experimental.UtilityClass;
